moved listings related db functions into seperate file
