+++ browse/services/listing/hybrid_listing.py
@@ -2,7 +2,7 @@ import logging
 
 from browse.services.listing.fs_listings import FsListingFilesService
 from browse.services.listing import YearCount
-from browse.services.database import get_yearly_article_counts
+from browse.services.database.listings import get_yearly_article_counts
 
 logger = logging.getLogger(__name__)
 logger.level = logging.DEBUG
